// core/packages/data/src/generated/fdevids/commodity.ts
export interface Commodity {
  id: number;
  symbol: string;
  category: string;
  name: string;
}
// ...
export const commoditiesBySymbol = new Map<string, Commodity>(
  commodities.map((r) => [r.symbol, r]),
);