// core/packages/data/src/generated/fdevids/engineers.ts
export interface Engineer {
  id: number;
  system_address: number | null;
  market_id: number | null;
  name: string;
}

export const engineers: Engineer[] = [
  {
    id: 300000,
    system_address: 3932277478114,
    market_id: 128673927,
    name: "Didi Vatermann",
  },
  {
    id: 300010,
    system_address: 1109989017963,
    market_id: 128674183,
    name: "Bill Turner",
  },
  {
    id: 300030,
    system_address: 4481966019282,
    market_id: 128674695,
    name: "Broo Tarquin",
  },
  {
    id: 300040,
    system_address: 2827992680811,
    market_id: 128674951,
    name: "The Sarge",
  },
  {
    id: 300050,
    system_address: 6131367744226,
    market_id: 128675207,
    name: "Zacariah Nemo",
  },
  {
    id: 300080,
    system_address: 1458309141194,
    market_id: 128675975,
    name: "Liz Ryder",
  },
  {
    id: 300090,
    system_address: 1733321102034,
    market_id: 128676231,
    name: "Hera Tani",
  },
  {
    id: 300100,
    system_address: 6681123623626,
    market_id: 128676487,
    name: "Felicity Farseer",
  },
  {
    id: 300110,
    system_address: 3790082132323,
    market_id: 128676743,
    name: "Ram Tah",
  },
  {
    id: 300120,
    system_address: 4305444669811,
    market_id: 128676999,
    name: "Lei Cheung",
  },
  {
    id: 300130,
    system_address: 12274907287851,
    market_id: 128677255,
    name: "Petra Olmanova",
  },
  {
    id: 300140,
    system_address: 10477373803,
    market_id: 128677511,
    name: "Colonel Bris Dekker",
  },
  {
    id: 300150,
    system_address: 48996147307082,
    market_id: 128677767,
    name: "Marsha Hicks",
  },
  {
    id: 300160,
    system_address: 3107241104074,
    market_id: 128678023,
    name: "Elvira Martuuk",
  },
  {
    id: 300180,
    system_address: 5031654888146,
    market_id: 128678535,
    name: "The Dweller",
  },
  {
    id: 300200,
    system_address: 121569805492,
    market_id: 128679047,
    name: "Marco Qwent",
  },
  {
    id: 300210,
    system_address: 24859942069665,
    market_id: 128679303,
    name: "Selene Jean",
  },
  {
    id: 300220,
    system_address: 113573366131,
    market_id: 128679559,
    name: "Professor Palin",
  },
  {
    id: 300230,
    system_address: 3932277478106,
    market_id: 128679815,
    name: "Lori Jameson",
  },
  {
    id: 300250,
    system_address: 4481899074282,
    market_id: 128680327,
    name: "Juri Ishmaak",
  },
  {
    id: 300260,
    system_address: 3107576681170,
    market_id: 128680583,
    name: "Tod 'The Blaster' McQuinn",
  },
  {
    id: 300270,
    system_address: 164098653,
    market_id: 128680839,
    name: "Tiana Fortune",
  },
  {
    id: 300280,
    system_address: 66038577537618,
    market_id: 128681095,
    name: "Mel Brandon",
  },
  {
    id: 300290,
    system_address: 11887629902418,
    market_id: 128681351,
    name: "Etienne Dorn",
  },
  {
    id: 300300,
    system_address: 594676730147,
    market_id: 128954244,
    name: "Chloe Sedesi",
  },
  {
    id: 400001,
    system_address: 7268024067513,
    market_id: 128972903,
    name: "Jude Navarro",
  },
  {
    id: 400002,
    system_address: 5068464399785,
    market_id: 128973159,
    name: "Domino Green",
  },
  {
    id: 400003,
    system_address: 7269634614689,
    market_id: 128973415,
    name: "Hero Ferrari",
  },
  {
    id: 400004,
    system_address: 2827975936355,
    market_id: 128973671,
    name: "Kit Fowler",
  },
  {
    id: 400005,
    system_address: 2832832893634,
    market_id: 128973927,
    name: "Wellington Beck",
  },
  {
    id: 400006,
    system_address: 3721329101171,
    market_id: 128974183,
    name: "Terra Velasquez",
  },
  {
    id: 400007,
    system_address: 16065190962585,
    market_id: 128974439,
    name: "Uma Laszlo",
  },
  {
    id: 400008,
    system_address: 8879744226018,
    market_id: 128974695,
    name: "Oden Geiger",
  },
  {
    id: 400009,
    system_address: 670686455169,
    market_id: 128974951,
    name: "Yarden Bond",
  },
  {
    id: 400010,
    system_address: 71536135676490,
    market_id: 128986843,
    name: "Baltanos",
  },
  {
    id: 400011,
    system_address: 38001031029322,
    market_id: 128987099,
    name: "Eleanor Bresa",
  },
  {
    id: 400012,
    system_address: 59166629864010,
    market_id: 128986587,
    name: "Rosa Dayette",
  },
  {
    id: 400013,
    system_address: 13736779007129,
    market_id: 128987355,
    name: "Yi Shen",
  },
];