// core/packages/data/src/generated/fdevids/systemallegiance.ts
// Auto-generated from FDevIDs
export interface SystemAllegiance {
  system_allegiance: string;
  name: string;
}

export const systemAllegiances: SystemAllegiance[] = [
  { system_allegiance: "Independent", name: "Independent" },
  { system_allegiance: "PilotsFederation", name: "Pilots Federation" },
  { system_allegiance: "Federation", name: "Federation" },
  { system_allegiance: "Empire", name: "Empire" },
  { system_allegiance: "Alliance", name: "Alliance" },
  { system_allegiance: "Guardian", name: "Guardian" },
  { system_allegiance: "Thargoid", name: "Thargoid" },
  { system_allegiance: "PlayerPilots", name: "Player Pilots" },
];